requirements.txt
# Web
edgartools
sec-edgar-downloader

# Data
pandas
packaging